Shamir's secret voting.
Instead of trusting the server with the votes, voters can register
as vote counters (before the vote begins). Then, every time a user
votes for a given schedule, every timeslot is encrypted using
Shamir's secret share scheme and each share is communicated secretly
to one of the vote counters. At the end of the voting process,
each vote counters add the encrypted shares and releases the sum,
which is then used by the server to calculate the final tally.

The architecture has changed substantially. Several new message
types have been introduced as well as synchronization between
server and client states.

Note that because of the vote secrecy, the users may lie and give
twenty votes to a given schedule. This should be solved using
zero-knowledge proofs.