Message initialization value and documentation update.
Old and/or incorrect references have been fixed in the
documentation comments.

A zero initialization was converted to a None initialization to
clearly signal the rest of the process if something goes wrong
(instead of just assuming a value of zero, which was mathematically
valid). The value corresponds to the x-coordinate of a vote counter.
In normal operation, it should never reach the initialization point
and still be unintialized in the end, but better defensive than
bug-hunting years later.

Note that the system would need something like a zero-knowledge
proof that votes are indeed in {0,1} for it to be generally valuable.
This is overkill for the intended case (scheduling meetings between
trustworthy people). If you need that I recommend checking out
Helios Voting. Alternatively, adding exponential ElGamal would be
absolutely welcomed.